# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gerenciamento Clinico",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/medics": {
            "get": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Listar médicos",
                "description": "Retorna uma lista de médicos",
                "operationId": "aa3b4ebbad7b186a5cf5a399230166f4",
                "responses": {
                    "200": {
                        "description": "Lista de médicos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Medic"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Criar médico",
                "description": "Cria um novo médico",
                "operationId": "08522a5ffeee36481ab47027e44fd209",
                "requestBody": {
                    "$ref": "#/components/requestBodies/StoreMedicRequest"
                },
                "responses": {
                    "201": {
                        "description": "Médico criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Medic"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/medics/{id}": {
            "get": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Exibir médico",
                "description": "Retorna os dados de um médico específico",
                "operationId": "ca82290fec96ee3f9cdbf49e4e6ea282",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do médico",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do médico",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Medic"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Atualizar médico",
                "description": "Atualiza os dados de um médico específico",
                "operationId": "c07a63b8664fdc8db77cf102ad0fefa5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do médico",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateMedicRequest"
                },
                "responses": {
                    "200": {
                        "description": "Médico atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Medic"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Excluir médico",
                "description": "Exclui um médico específico",
                "operationId": "d19d4df17f6aa85a05ebf2fe2b2aa752",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do médico",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Médico excluído"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Medic": {
                "title": "Médico",
                "description": "Modelo de Médico",
                "required": [
                    "name",
                    "email",
                    "crm"
                ],
                "properties": {
                    "id": {
                        "description": "ID do médico",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nome do médico",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email do médico",
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "description": "Telefone do médico",
                        "type": "string"
                    },
                    "speciality": {
                        "description": "Especialidade do médico",
                        "type": "string"
                    },
                    "crm": {
                        "description": "CRM do médico",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
...
            "StoreMedicRequest": {
                "description": "Dados para criar um médico",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Nome do médico",
                                    "type": "string"
                                },
                                "email": {
                                    "description": "Email do médico",
                                    "type": "string",
                                    "format": "email"
                                },
                                "phone": {
                                    "description": "Telefone do médico",
                                    "type": "string"
                                },
                                "speciality": {
                                    "description": "Especialidade do médico",
                                    "type": "string"
                                },
                                "crm": {
                                    "description": "CRM do médico",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "UpdateMedicRequest": {
                "description": "Dados para atualizar um médico",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Nome do médico",
                                    "type": "string"
                                },
                                "email": {
                                    "description": "Email do médico",
                                    "type": "string",
                                    "format": "email"
                                },
                                "phone": {
                                    "description": "Telefone do médico",
                                    "type": "string"
                                },
                                "speciality": {
                                    "description": "Especialidade do médico",
                                    "type": "string"
                                },
                                "crm": {
                                    "description": "CRM do médico",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}